+++ examples/for-tests-react-16/src/AppWithReactDomRouter.js
@@ -26,8 +26,8 @@ if (authRecipe === "thirdparty") {
 } else if (authRecipe === "both") {
     routesRenderer = (router) => {
         return [
-            ...ThirdPartyPreBuiltUI.getReactRouterDomRoutes(router),
             ...EmailPasswordPreBuiltUI.getReactRouterDomRoutes(router),
+            ...ThirdPartyPreBuiltUI.getReactRouterDomRoutes(router),
         ];
     };
 } else if (authRecipe === "thirdpartyemailpassword") {