+++ examples/with-next-iframe/config/backendConfig.js
@@ -1,6 +1,7 @@
 import EmailPassword from "supertokens-node/recipe/emailpassword";
 import EmailVerification from "supertokens-node/recipe/emailverification";
 import Session from "supertokens-node/recipe/session";
+import Dashboard from "supertokens-node/recipe/dashboard";
 
 import { appInfo } from "./appInfo";
 
@@ -20,6 +21,7 @@ export const backendConfig = () => {
                 cookieSameSite: "none",
                 antiCsrf: "VIA_TOKEN",
             }),
+            Dashboard.init(),
         ],
         isInServerlessEnv: true,
     };